license-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
slowapi==0.1.9
python-multipart==0.0.20
